pyproject.toml
[project]
name = "assignment-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Unazabutt", email = "unazabutt44@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
assignment-3 = "assignment_3:main"
start_agent = "assignment_3.my_first_agent:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
